+++ java/connectors/semantickernel-connectors-memory-sqlite/src/test/java/com/microsoft/semantickernel/connectors/memory/sqlite/SQLiteMemoryStoreTest.java
@@ -397,7 +397,7 @@ public class SQLiteMemoryStoreTest {
         this._db.upsertAsync(collection, testRecord).block();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 this._db
                         .getNearestMatchesAsync(
@@ -483,7 +483,7 @@ public class SQLiteMemoryStoreTest {
         this._db.upsertAsync(collection, testRecord).block();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 this._db
                         .getNearestMatchesAsync(
@@ -563,7 +563,7 @@ public class SQLiteMemoryStoreTest {
         this._db.upsertAsync(collection, testRecord).block();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 this._db
                         .getNearestMatchesAsync(collection, compareEmbedding, 0, threshold, false)
@@ -583,7 +583,7 @@ public class SQLiteMemoryStoreTest {
         this._db.createCollectionAsync(collection).block();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 this._db
                         .getNearestMatchesAsync(
@@ -664,7 +664,7 @@ public class SQLiteMemoryStoreTest {
         this._db.upsertAsync(collection, testRecord).block();
 
         // Act
-        double threshold = 0.75;
+        float threshold = 0.75f;
         Tuple2<MemoryRecord, Float> topNResultDefault =
                 this._db
                         .getNearestMatchAsync(collection, compareEmbedding, threshold, false)
@@ -754,7 +754,7 @@ public class SQLiteMemoryStoreTest {
         this._db.upsertAsync(collection, testRecord).block();
 
         // Act
-        double threshold = 0.75;
+        float threshold = 0.75f;
         Tuple2<MemoryRecord, Float> topNResult =
                 this._db
                         .getNearestMatchAsync(collection, compareEmbedding, threshold, false)
@@ -776,7 +776,7 @@ public class SQLiteMemoryStoreTest {
         this._db.createCollectionAsync(collection).block();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Tuple2<MemoryRecord, Float> topNResults =
                 this._db
                         .getNearestMatchAsync(collection, compareEmbedding, threshold, false)
@@ -811,7 +811,7 @@ public class SQLiteMemoryStoreTest {
         // Act
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 this._db
-                        .getNearestMatchesAsync(collection, compareEmbedding, topN, 0.75, true)
+                        .getNearestMatchesAsync(collection, compareEmbedding, topN, 0.75f, true)
                         .block();
         Collection<String> topNKeys =
                 topNResults.stream()
